+++ src/main/java/es/unican/atc/memory/MemContents.java
@@ -54,7 +54,7 @@ class MemContents implements Cloneable, HexModel {
     }
 
     private void fireBytesChanged(long start, long numBytes,
-            int[] oldValues) {
+            long[] oldValues) {
         if (listeners == null) return;
         boolean found = false;
         for (HexModelListener l : listeners) {
@@ -85,7 +85,7 @@ class MemContents implements Cloneable, HexModel {
     public int getLogLength() { return addrBits; }
     public int getWidth() { return width; }
 
-    public int get(long addr) {
+    public long get(long addr) {
         int page = (int) (addr >>> PAGE_SIZE_BITS);
         int offs = (int) (addr & PAGE_MASK);
         if (page < 0 || page >= pages.length || pages[page] == null) return 0;
@@ -104,21 +104,21 @@ class MemContents implements Cloneable, HexModel {
         return true;
     }
 
-    public void set(long addr, int value) {
+    public void set(long addr, long value) {
         int page = (int) (addr >>> PAGE_SIZE_BITS);
         int offs = (int) (addr & PAGE_MASK);
-        int old = pages[page] == null ? 0 : pages[page].get(offs) & mask;
-        int val = value & mask;
+        long old = pages[page] == null ? 0 : pages[page].get(offs) & mask;
+        long val = value & mask;
         if (old != val) {
             if (pages[page] == null) {
                 pages[page] = MemContentsSub.createContents(PAGE_SIZE, width);
             }
             pages[page].set(offs, val);
-            fireBytesChanged(addr, 1, new int[] { old });
+            fireBytesChanged(addr, 1, new long[] { old });
         }
     }
 
-    public void set(long start, int[] values) {
+    public void set(long start, long[] values) {
         if (values.length == 0) return;
 
         int pageStart = (int) (start >>> PAGE_SIZE_BITS);
@@ -130,7 +130,7 @@ class MemContents implements Cloneable, HexModel {
             ensurePage(pageStart);
             MemContentsSub.ContentsInterface page = pages[pageStart];
             if (!page.matches(values, startOffs, mask)) {
-                int[] oldValues = page.get(startOffs, values.length);
+                long[] oldValues = page.get(startOffs, values.length);
                 page.load(startOffs, values, mask);
                 if (page.isClear()) pages[pageStart] = null;
                 fireBytesChanged(start, values.length, oldValues);
@@ -142,18 +142,18 @@ class MemContents implements Cloneable, HexModel {
                 nextOffs = 0;
             } else {
                 ensurePage(pageStart);
-                int[] vals = new int[PAGE_SIZE - startOffs];
+                long[] vals = new long[PAGE_SIZE - startOffs];
                 System.arraycopy(values, 0, vals, 0, vals.length);
                 MemContentsSub.ContentsInterface page = pages[pageStart];
                 if (!page.matches(vals, startOffs, mask)) {
-                    int[] oldValues = page.get(startOffs, vals.length);
+                    long[] oldValues = page.get(startOffs, vals.length);
                     page.load(startOffs, vals, mask);
                     if (page.isClear()) pages[pageStart] = null;
                     fireBytesChanged(start, PAGE_SIZE - pageStart, oldValues);
                 }
                 nextOffs = vals.length;
             }
-            int[] vals = new int[PAGE_SIZE];
+            long[] vals = new long[PAGE_SIZE];
             int offs = nextOffs;
             for (int i = pageStart + 1; i < pageEnd; i++, offs += PAGE_SIZE) {
                 MemContentsSub.ContentsInterface page = pages[i];
@@ -170,7 +170,7 @@ class MemContents implements Cloneable, HexModel {
                 if (page != null) {
                     System.arraycopy(values, offs, vals, 0, PAGE_SIZE);
                     if (!page.matches(vals, startOffs, mask)) {
-                        int[] oldValues = page.get(0, PAGE_SIZE);
+                        long[] oldValues = page.get(0, PAGE_SIZE);
                         page.load(0, vals, mask);
                         if (page.isClear()) pages[i] = null;
                         fireBytesChanged((long) i << PAGE_SIZE_BITS, PAGE_SIZE, oldValues);
@@ -179,11 +179,11 @@ class MemContents implements Cloneable, HexModel {
             }
             if (endOffs >= 0) {
                 ensurePage(pageEnd);
-                vals = new int[endOffs + 1];
+                vals = new long[endOffs + 1];
                 System.arraycopy(values, offs, vals, 0, endOffs + 1);
                 MemContentsSub.ContentsInterface page = pages[pageEnd];
                 if (!page.matches(vals, startOffs, mask)) {
-                    int[] oldValues = page.get(0, endOffs + 1);
+                    long[] oldValues = page.get(0, endOffs + 1);
                     page.load(0, vals, mask);
                     if (page.isClear()) pages[pageEnd] = null;
                     fireBytesChanged((long) pageEnd << PAGE_SIZE_BITS, endOffs + 1, oldValues);
@@ -192,7 +192,7 @@ class MemContents implements Cloneable, HexModel {
         }
     }
 
-    public void fill(long start, long len, int value) {
+    public void fill(long start, long len, long value) {
         if (len == 0) return;
 
         int pageStart = (int) (start >>> PAGE_SIZE_BITS);
@@ -203,11 +203,11 @@ class MemContents implements Cloneable, HexModel {
 
         if (pageStart == pageEnd) {
             ensurePage(pageStart);
-            int[] vals = new int[(int) len];
+            long[] vals = new long[(int) len];
             Arrays.fill(vals, value);
             MemContentsSub.ContentsInterface page = pages[pageStart];
             if (!page.matches(vals, startOffs, mask)) {
-                int[] oldValues = page.get(startOffs, (int) len);
+                long[] oldValues = page.get(startOffs, (int) len);
                 page.load(startOffs, vals, mask);
                 if (value == 0 && page.isClear()) pages[pageStart] = null;
                 fireBytesChanged(start, len, oldValues);
@@ -220,11 +220,11 @@ class MemContents implements Cloneable, HexModel {
                     // nothing to do
                 } else {
                     ensurePage(pageStart);
-                    int[] vals = new int[PAGE_SIZE - startOffs];
+                    long[] vals = new long[PAGE_SIZE - startOffs];
                     Arrays.fill(vals, value);
                     MemContentsSub.ContentsInterface page = pages[pageStart];
                     if (!page.matches(vals, startOffs, mask)) {
-                        int[] oldValues = page.get(startOffs, vals.length);
+                        long[] oldValues = page.get(startOffs, vals.length);
                         page.load(startOffs, vals, mask);
                         if (value == 0 && page.isClear()) pages[pageStart] = null;
                         fireBytesChanged(start, PAGE_SIZE - pageStart, oldValues);
@@ -236,13 +236,13 @@ class MemContents implements Cloneable, HexModel {
                     if (pages[i] != null) clearPage(i);
                 }
             } else {
-                int[] vals = new int[PAGE_SIZE];
+                long[] vals = new long[PAGE_SIZE];
                 Arrays.fill(vals, value);
                 for (int i = pageStart + 1; i < pageEnd; i++) {
                     ensurePage(i);
                     MemContentsSub.ContentsInterface page = pages[i];
                     if (!page.matches(vals, 0, mask)) {
-                        int[] oldValues = page.get(0, PAGE_SIZE);
+                        long[] oldValues = page.get(0, PAGE_SIZE);
                         page.load(0, vals, mask);
                         fireBytesChanged((long) i << PAGE_SIZE_BITS, PAGE_SIZE, oldValues);
                     }
@@ -254,10 +254,10 @@ class MemContents implements Cloneable, HexModel {
                     // nothing to do
                 } else {
                     ensurePage(pageEnd);
-                    int[] vals = new int[endOffs + 1];
+                    long[] vals = new long[endOffs + 1];
                     Arrays.fill(vals, value);
                     if (!page.matches(vals, 0, mask)) {
-                        int[] oldValues = page.get(0, endOffs + 1);
+                        long[] oldValues = page.get(0, endOffs + 1);
                         page.load(0, vals, mask);
                         if (value == 0 && page.isClear()) pages[pageEnd] = null;
                         fireBytesChanged((long) pageEnd << PAGE_SIZE_BITS, endOffs + 1, oldValues);
@@ -277,10 +277,10 @@ class MemContents implements Cloneable, HexModel {
 
     private void clearPage(int index) {
         MemContentsSub.ContentsInterface page = pages[index];
-        int[] oldValues = new int[page.getLength()];
+        long[] oldValues = new long[page.getLength()];
         boolean changed = false;
         for (int j = 0; j < oldValues.length; j++) {
-            int val = page.get(j) & mask;
+            long val = page.get(j) & mask;
             oldValues[j] = val;
             if (val != 0) changed = true;
         }
